backend/code/requirements.txt
amqp==2.6.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kombu==4.6.11
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
twilio==7.16.1
tzdata==2022.7
urllib3==1.26.14
uuid==1.30
vine==1.3.0
django-braintree
paypalrestsdk
psycopg2-binary>=2.8
django-redis